refactor(api): git repositories 라우트를 전용 모듈로 추출
5개 라우트를 api/git/repositories.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 동일.

_legacy.py에서 이 5개 핸들러만 쓰던 models.git 심볼 9개의 import를 제거했다
(GitRepository·GitRepositoryCreate·GitRepositoryListResponse·GitRepositoryUpdate·
list_git_repositories·register_git_repository·get_git_repository·
update_git_repository·delete_git_repository). 자기 변경이 만든 미사용 import만
정리한 것이며, 그 외 코드는 손대지 않았다.
